feat(serverHandler): prevent mutate items under directory if aliased

+++ src/serverHandler/delete.go
@@ -4,10 +4,13 @@ import (
 	"os"
 )
 
-func (h *handler) deleteItems(fsPrefix string, files []string) {
+func (h *handler) deleteItems(fsPrefix string, files []string, aliasSubItems []os.FileInfo) {
 	errs := []error{}
 
 	for _, filename := range files {
+		if containsItem(aliasSubItems, filename) {
+			continue
+		}
 		err := os.RemoveAll(fsPrefix + "/" + filename)
 		if err != nil {
 			errs = append(errs, err)
